# packages/ar-auth/openapi/auth.openapi.yaml
openapi: 3.2.0
info:
  title: Authrim Authentication API
  version: 0.3.3
  summary: Browser and API authentication endpoints.
  description: >
    Initial OpenAPI contract for endpoints owned by `@authrim/ar-auth`. This document intentionally
    contains route contracts only; do not add secrets, private test credentials, or vulnerability
    notes.
  x-authrim-coverage: initial route inventory
servers:
  - url: https://{tenant_domain}
    variables:
      tenant_domain:
        default: auth.example.com
...
paths:
...
  /api/auth/directory-password/migration/passkey/options:
    post:
      tags:
        - Directory Password
      summary: Create WebAuthn registration options for a directory password migration transaction.
      requestBody:
        $ref: '#/components/requestBodies/DirectoryMigrationPasskeyOptions'
      responses:
        '200':
          $ref: '#/components/responses/Any'
        '400':
          $ref: '#/components/responses/Any'
        '401':
          $ref: '#/components/responses/Any'
        '403':
          $ref: '#/components/responses/Any'
...
  /api/admin/auth/passkey/options:
    post:
      tags:
        - Auth
      operationId: postApiAdminAuthPasskeyOptions
      summary: Create Admin passkey login options.
      description:
        Creates discoverable WebAuthn authentication options for Admin UI passkey login and
        stores a short-lived challenge.
      responses:
        '200':
          $ref: '#/components/responses/AdminPasskeyOptions'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
      requestBody:
        $ref: '#/components/requestBodies/AdminPasskeyOptions'
...
  /api/admin/setup-token/passkey/complete:
    post:
      tags:
        - Auth
      operationId: postApiAdminSetupTokenPasskeyComplete
      summary: Complete Admin UI setup passkey registration.
      description:
        Verifies the setup-token passkey registration response, stores the admin passkey, and
        marks the setup token as used.
      responses:
        '200':
          $ref: '#/components/responses/AdminPasskeyVerify'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
      requestBody:
        $ref: '#/components/requestBodies/AdminPasskeyVerify'
  /api/admin/setup-token/passkey/options:
    post:
      tags:
        - Auth
      operationId: postApiAdminSetupTokenPasskeyOptions
      summary: Create setup-token WebAuthn registration options.
      description:
        Validates an Admin UI setup token and returns WebAuthn registration options plus a
        challenge identifier.
      responses:
        '200':
          $ref: '#/components/responses/AdminPasskeyOptions'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
      requestBody:
        $ref: '#/components/requestBodies/AdminPasskeyOptions'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: authrim_session
...
  requestBodies:
...
    DirectoryMigrationPasskeyOptions:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/DirectoryMigrationPasskeyOptionsRequest'
          example:
            transaction_id: damt_123
            transaction_token: migration-transaction-token
            display_name: Example User
...
    AdminPasskeyOptions:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AdminPasskeyOptionsRequest'
          example:
            token: admin_setup_token_123
            rp_id: admin.example.com
    AdminPasskeyVerify:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AdminPasskeyVerifyRequest'
          example:
            token: admin_setup_token_123
            challenge_id: challenge_123
            passkey_response:
              id: credential-id
              type: public-key
              response: {}
            origin: https://admin.example.com
...
  responses:
...
    Any:
      description: Successful response. The representation can be JSON, HTML, or a redirect.
    Error:
      description: Error response.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    AdminPasskeyOptions:
      description: Successful JSON response.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AdminPasskeyOptionsResponse'
          example:
            options:
              challenge: challenge
              rp:
                name: Authrim
                id: admin.example.com
              user:
                id: admin_123
                name: admin@example.com
                displayName: Admin User
            challenge_id: challenge_123
    AdminPasskeyVerify:
      description: Successful JSON response.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AdminPasskeyVerifyResponse'
          example:
            success: true
            user:
              id: admin_123
              email: admin@example.com
              name: Admin User
...
  schemas:
...
    Error:
      type: object
      additionalProperties: true
      properties:
        error:
          type: string
        error_description:
          type: string
        message:
          type: string
        webauthn_signal:
          $ref: '#/components/schemas/WebAuthnSignalHint'
    WebAuthnSignalHint:
      type: object
      description:
        Optional browser-side WebAuthn Signal API hint. When unknown_credential is true, clients
        that just received a WebAuthn credential assertion may call
        PublicKeyCredential.signalUnknownCredential() for that credential ID.
      properties:
        unknown_credential:
          type: boolean
...
    WebAuthnOptions:
      type: object
      additionalProperties: true
      description: WebAuthn options generated by @simplewebauthn/server.
    WebAuthnCredential:
      type: object
      additionalProperties: true
      description: WebAuthn credential response from the browser.
...
    DirectoryMigrationPasskeyOptionsRequest:
      type: object
      additionalProperties: false
      properties:
        transaction_id:
          type: string
          minLength: 1
          example: damt_123
        transaction_token:
          type: string
          minLength: 1
          example: migration-transaction-token
        display_name:
          type: string
          maxLength: 160
          example: Example User
      required:
        - transaction_id
        - transaction_token
...
    AdminPasskeyOptionsRequest:
      type: object
      properties:
        token:
          type: string
          example: admin_setup_token_123
        rp_id:
          type: string
          example: admin.example.com
      required:
        - token
        - rp_id
    AdminPasskeyOptionsResponse:
      type: object
      properties:
        options:
          $ref: '#/components/schemas/WebAuthnOptions'
        challenge_id:
          type: string
          example: challenge_123
      required:
        - options
        - challenge_id
    AdminPasskeyVerifyRequest:
      type: object
      properties:
        token:
          type: string
          example: admin_setup_token_123
        challenge_id:
          type: string
          example: challenge_123
        passkey_response:
          $ref: '#/components/schemas/WebAuthnCredential'
        origin:
          type: string
          format: uri
          example: https://admin.example.com
      required:
        - token
        - challenge_id
        - passkey_response
        - origin
    AdminPasskeyVerifyResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        user:
          type: object
          properties:
            id:
              type: string
              example: admin_123
            email:
              type: string
              format: email
              example: admin@example.com
            name:
              anyOf:
                - type: string
                  example: Admin User
                - type: 'null'
      required:
        - success
        - user
...